requirements.txt
pyyaml>=6.0
jsonschema>=4.0.0
opencv-python>=4.6.0
numpy>=1.22.0 